services/api-gateway/requirements.txt
# API Gateway dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
clickhouse-connect==0.6.23
aiokafka==0.10.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
asyncpg==0.29.0
httpx==0.26.0
